refactor(orm): extract audit persistence
