+++ tests/Headless.AuditLog.EntityFramework.Tests.Integration/AuditLogIntegrationTests.cs
@@ -90,6 +90,38 @@ public sealed class AuditLogIntegrationTests : TestBase
     }
 
     [Fact]
+    public async Task sync_save_changes_false_with_automatic_audit_preserves_entity_state_and_persists_audit_once()
+    {
+        // given
+        var (sp, conn) = await AuditIntegrationFixture.CreateAsync();
+        await using var _ = conn;
+        await using var __ = sp;
+        await using var scope = sp.CreateAsyncScope();
+        await using var db = scope.ServiceProvider.GetRequiredService<AuditTestDbContext>();
+
+        var order = new GeneratedOrder { CustomerName = "Generated" };
+        db.GeneratedOrders.Add(order);
+
+        // when
+#pragma warning disable CA1849 // Call async methods when in an async method
+        db.SaveChanges(acceptAllChangesOnSuccess: false);
+#pragma warning restore CA1849 // Call async methods when in an async method
+
+        // then
+        order.Id.Should().BePositive();
+        db.Entry(order).State.Should().Be(EntityState.Added);
+        db.ChangeTracker.Entries<AuditLogEntry>().Should().BeEmpty();
+
+        var entries = await db.Set<AuditLogEntry>().AsNoTracking().ToListAsync(AbortToken);
+        entries.Should().ContainSingle();
+
+        var entry = entries[0];
+        entry.Action.Should().Be(AuditActionNames.Created);
+        entry.EntityType.Should().Be(typeof(GeneratedOrder).FullName);
+        entry.EntityId.Should().Be(order.Id.ToString(CultureInfo.InvariantCulture));
+    }
+
+    [Fact]
     public async Task read_audit_log_query_returns_dto_results()
     {
         // given
